Eliminated some compiler warnings, fixing a couple of small bugs in the process.
git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@13144 67ed7778-7388-44ab-90cf-0a291f65f57c

+++ camlibs/kodak/ez200/ez200.c
@@ -261,7 +261,7 @@ get_file_func (CameraFilesystem *fs, const char *folder, const char *filename,
 	if (!data) return GP_ERROR_NO_MEMORY;
 
 	data_start = data + (HEADER_SIZE - DATA_HEADER_SIZE);
-	GP_DEBUG("data - data_start : %p %p : %lx",data, data_start, data_start - data);
+	GP_DEBUG("data - data_start : %p %p : %lx",data, data_start, (long) (data_start - data));
 
     	ez200_read_picture_data   (camera->port, data_start, len, n);
 	ez200_read_picture_header (camera->port, data);
